# Add additional support for array indexing in Set