@JoshKCarroll 2017-09-08
Should we include a test to note this behavior (which maybe isn't ideal but I think that's ok for now)?

```
	{
		desc:    "set indexed path to object on empty JSON, index value is ignored",
		json:    `{}`,
		path:    []string{"top", "[0]", "middle", "[1]", "bottom"},
		setData: `"value"`,
		isFound: true,
		data:    `{"top":[{"middle":[{"bottom":"value"}]}]}`,
	},
```