@chilland 2017-09-08
@JoshKCarroll I think thats a good call. As far as the right behavior, I'm not entirely sure what the alternative is. Throw an error?