@JoshKCarroll 2017-09-08
Also, if I understand the code correctly, this test would pass.. when I think the expected behavior would be to append rather than replace the array. Maybe that's ok right now but should file an issue if we don't want to fix it. Or am I misunderstanding?
```
	{
		desc:    "array is replaced instead of expected append",
		json:    `{"top":[{"middle":[]}]}`,
		path:    []string{"top", "[1]", "second", "[0]"},
		setData: `"value"`,
		isFound: true,
		data:    `{"top":[{"second":["value"]}]},
        // Expected: `{"top":[{"middle":[]}, {"second":["value"]}]}
	},
```