@JoshKCarroll 2017-09-08
Yeah I think either a nil-padded array or throw an error