@JoshKCarroll 2017-09-13
Taking over for Casey and continuing to work on this at https://github.com/qntfy/jsonparser/pull/2